services/ai-chat/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-generativeai==0.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
PyJWT==2.8.0
